requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
yt-dlp>=2024.4.9
Jinja2>=3.1
isodate>=0.6
requests>=2.31
python-dotenv>=1.0
